refactor(OutputView): LineType을 내부 enum 클래스로 이동
